{"schema":"decodex/commit/1","summary":"Split external adapter validation modules","authority":"manual"}
